requirements.txt
astroid==2.4.1
autopep8==1.5.2
certifi==2020.4.5.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pkg-resources==0.0.0
pycodestyle==2.6.0
pylint==2.5.2
pymongo==3.10.1
PySocks==1.7.1
requests-oauthlib==1.3.0
toml==0.10.1
tweepy==3.8.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
beautifulsoup4==4.8.2
chardet==3.0.4
Click==7.0
cssselect==1.1.0
decorator==4.4.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
idna==2.9
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.18.1
Pillow==7.0.0
python-dateutil==2.8.1
python-dotenv==0.11.0
PyYAML==5.3
requests==2.23.0
requests-file==1.4.3
Sastrawi==1.0.1
six==1.14.0
soupsieve==2.0
tinysegmenter==0.3
tldextract==2.2.2
validators==0.14.2
Werkzeug==1.0.0
grpcio
protobuf
pandas
tqdm
sklearn